requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pandas>=1.5.0
matplotlib>=3.6.0
Pillow>=9.4.0
numpy>=1.23.0
imageio>=2.28.1
tqdm>=4.65.0
lpips>=0.1.4
